+++ Games/Games.Risk.Application/ComputerPlayerOptions/AttackOption.cs
@@ -4,6 +4,6 @@
     {
         public int IndexOfTerritoryWhereAttackOriginates { get; set; }
         public int IndexOfTerritoryUnderAttack { get; set; }
-        public double OpportunityRating { get; set; }
+        public int OpportunityRating { get; set; }
     }
 }
